requirements.txt
git+https://github.com/Rapptz/discord.py@master
git+https://github.com/Rapptz/discord-ext-menus

yarl
aiohttp
humanize
python-dateutil
tabulate
Pillow
uwuify
psutil
asyncpg
PyYAML
async-cse
discord-flags
async-timeout
googletrans
import-expression

